requirements.txt
fastapi
uvicorn
streamlit
google-generativeai
groq
python-dotenv
pandas
plotly
requests
pydantic
